pyproject.toml
[tool.poetry]
name = "beetle-backups"
version = "0.1.0"
description = ""
authors = ["Gilson Paulino <gilsonbp@gmail.com>"]
readme = "README.md"
packages = [{include = "beetle_backups"}]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.26.37"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
newrelic = "^8.5.0"
asgi-correlation-id = "^3.2.1"
python-slugify = "^7.0.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
pre-commit = "^2.21.0"
mypy = "^0.991"
ipython = "^8.7.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"


[tool.mypy]
plugins = "pydantic.mypy"
ignore_missing_imports = true
namespace_packages = true
